+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### master
 - enable "tmux auto start" for OS X
+- enable customizing "tmux auto start" for OS X
 
 ### v2.0.0, 2015-02-15
 - enable automatic environment restore when tmux is started