+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### master
 - enable "tmux auto start" for OS X
 - enable customizing "tmux auto start" for OS X
+- fix errors when creating a launchd plist file for auto-start on OS X
 
 ### v2.0.0, 2015-02-15
 - enable automatic environment restore when tmux is started